+++ tests/genai/judges/test_judge_utils.py
@@ -1056,7 +1056,10 @@ def test_invoke_judge_model_databricks_success_not_in_databricks(
         feedback = invoke_judge_model(**kwargs)
 
         mock_invoke_db.assert_called_once_with(
-            model_name=expected_model_name, prompt="Test prompt", num_retries=10
+            model_name=expected_model_name,
+            prompt="Test prompt",
+            num_retries=10,
+            response_format=None,
         )
         mock_in_db.assert_called_once()
 
@@ -1110,7 +1113,10 @@ def test_invoke_judge_model_databricks_success_in_databricks(
             num_completion_tokens=8,
         )
         mock_invoke_db.assert_called_once_with(
-            model_name=expected_model_name, prompt="Test prompt", num_retries=10
+            model_name=expected_model_name,
+            prompt="Test prompt",
+            num_retries=10,
+            response_format=None,
         )
         mock_in_db.assert_called_once()
 
@@ -1142,7 +1148,7 @@ def test_invoke_judge_model_databricks_source_id(model_uri: str) -> None:
         "test-model" if model_uri.startswith("databricks") else "databricks-gpt-oss-120b"
     )
     mock_invoke_db.assert_called_once_with(
-        model_name=expected_model_name, prompt="Test prompt", num_retries=10
+        model_name=expected_model_name, prompt="Test prompt", num_retries=10, response_format=None
     )
     assert feedback.source.source_id == f"databricks:/{expected_model_name}"
 
@@ -1185,7 +1191,10 @@ def test_invoke_judge_model_databricks_failure_in_databricks(
             error_message=mock.ANY,  # Check that error message contains the exception
         )
         mock_invoke_db.assert_called_once_with(
-            model_name=expected_model_name, prompt="Test prompt", num_retries=10
+            model_name=expected_model_name,
+            prompt="Test prompt",
+            num_retries=10,
+            response_format=None,
         )
         mock_in_db.assert_called_once()
 
@@ -1238,7 +1247,10 @@ def test_invoke_judge_model_databricks_telemetry_error_handling(
             num_completion_tokens=3,
         )
         mock_invoke_db.assert_called_once_with(
-            model_name=expected_model_name, prompt="Test prompt", num_retries=10
+            model_name=expected_model_name,
+            prompt="Test prompt",
+            num_retries=10,
+            response_format=None,
         )
         mock_in_db.assert_called_once()
